rename util files for epub
