+++ src/screens/MediaSettings/Utils/epub/chapter-navigator.js
@@ -1,11 +1,11 @@
 import { util } from 'utils';
-import { epubState } from './setstate';
+import { epubState } from './epub-state';
 import { 
     NAV_CLOSE,
     NAV_HIDE,
     NAV_SHOW,
     SUB_CHAPTER_ID_PREFIX
-} from './constants'
+} from './constants';
 
 const { elem } = util;
