+++ src/screens/MediaSettings/Utils/epub/epub-downloader.js
@@ -1,10 +1,10 @@
 import _ from 'lodash';
 import { prompt, CTEpubGenerator } from 'utils';
-import { EDITOR_MARKDOWN } from './constants';
 import { setup } from '../setup';
 import { getImageUrl } from './util';
 import { markdown2HTML, parseText } from './chapter-html-converter';
-import { epubState } from './setstate';
+import { epubState } from './epub-state';
+import { EDITOR_MARKDOWN } from './constants';
 
 
 function formatEpubChapter(chapter) {
