+++ src/screens/MediaSettings/Utils/epub/index.js
@@ -1,24 +1,28 @@
-import { epubState } from './setstate';
+import { epubState } from '././epub-state';
+
 import * as epubConstants from './constants';
 import * as epubUtils from './util';
-import * as chapterEditor from './chapter-editor';
-import * as chaperSplitter from './chapter-spiltter';
-import * as downloader from './downloader';
 import * as chapterHtmlConverter from './chapter-html-converter';
 import * as chapterNavigator from './chapter-navigator';
-
-export { connectWithRedux, epubStore, epubContext } from 'redux/media-settings/epub'
+import * as chaperSplitter from './chapter-spiltter';
+import * as chapterEditor from './chapter-editor';
+import * as epubDownloader from './epub-downloader';
 
 export const epub = {
     state: epubState,
-    
     resetEpubData: () => epubState.setEpubData([]),
 
     ...epubConstants,
     ...epubUtils,
-    ...chapterEditor,
-    ...chaperSplitter,
-    ...downloader,
     ...chapterHtmlConverter,
-    ...chapterNavigator
+    ...chapterNavigator,
+    ...chaperSplitter,
+    ...chapterEditor,
+    ...epubDownloader
 };
+
+export { 
+    epubStore, 
+    epubContext, 
+    connectWithRedux 
+} from 'redux/media-settings/epub';
